Revert "Cambio en vistas y consultar si gané una subasta"
This reverts commit bd6e2f7d4244dcc884592d86f9bf2ea3f25cb973.

+++ Bestnid/db/migrate/20141110150424_create_oferta.rb
@@ -2,7 +2,7 @@ class CreateOferta < ActiveRecord::Migration
   def change
     create_table :oferta do |t|
       t.text :motivo
-      t.float :monto
+      t.integer :monto
       t.references :usuario, index: true
       t.references :producto, index: true
       t.timestamps
